feat(schemas): float 필드를 Decimal로 변경하여 정밀도 향상 및 JSON 직렬화 개선

+++ backend/schemas/admin.py
@@ -1,8 +1,9 @@
 from pydantic import BaseModel, Field
+from decimal import Decimal
 
 class FeeUpdate(BaseModel):
-    fee_rate: float = Field(..., ge=0.0, le=1.0, description="Trading fee rate (e.g., 0.001 for 0.1%)")
+    fee_rate: Decimal = Field(..., ge=0.0, le=1.0, description="Trading fee rate (e.g., 0.001 for 0.1%)")
 
 class PriceUpdate(BaseModel):
     ticker_id: str
-    price: float
+    price: Decimal
